Compilable version of FVSsn
git-svn-id: svn+ssh://svn.code.sf.net/p/open-fvs/code/trunk@54 90b89b48-9230-0de4-5065-812dd12741b8

+++ fire/sn/src/fmcba.f
@@ -141,7 +141,7 @@ C-----------
         ELSE
           SHRUBAGE = (IYR - IY(1) + 5) ! so default is 5 years before inventory
         ENDIF
-        SHRUBAGE = MIN(20,MAX(1,INT(SHRUBAGE)))
+        SHRUBAGE = MIN(20.0,MAX(1.0,SHRUBAGE))
  
         DO L=1,8
           FULIV3(L) = FULIV2(L,J)
